(PC-28265)[PRO] refactor: standardize formik form
