+++ pro/src/pages/VenueEdition/VenueEditionFormScreen.module.scss
@@ -1,10 +1,6 @@
 @use "styles/mixins/_rem.scss" as rem;
 @use "styles/variables/_colors.scss" as colors;
 
-.venue-form {
-  margin-top: rem.torem(32px);
-}
-
 .page-status {
   margin-bottom: rem.torem(32px);
 }
